TP2/requirements.txt
aiohttp>=3.9.0
beautifulsoup4>=4.12.2
lxml>=4.9.3
Pillow>=10.4.0
selenium>=4.24.0
aiofiles>=24.1.0
playwright>=1.47.0

